refactor(ports): extract TTSEngine ABC to ports/tts.py (ADR-059 QW3)
- Add src/voicecli/ports/__init__.py and src/voicecli/ports/tts.py with
  the TTSEngine ABC only — no torch, no CUDA, no infrastructure imports
- engine.py: remove inline ABC definition, import TTSEngine from ports.tts
  (re-exports it so existing callers via voicecli.engine remain unbroken)
